Extracts pagination helpers to a separate target
